Move numeric comparison functions into a submodule
And document them.
