+++ src/builtin/functions/functions.rs
@@ -13,7 +13,6 @@ use crate::TulispObject;
 use crate::TulispValue;
 use crate::{destruct_bind, list};
 use std::convert::TryInto;
-use std::rc::Rc;
 use tulisp_proc_macros::{crate_fn, crate_fn_no_eval};
 
 pub(super) fn reduce_with(
@@ -105,37 +104,6 @@ fn mark_tail_calls(
 }
 
 pub(crate) fn add(ctx: &mut TulispContext) {
-    // // TODO: >, >=, <, <= - need to be able to support more than 2 args
-    fn gt(ctx: &mut TulispContext, rest: &TulispObject) -> Result<TulispObject, Error> {
-        reduce_with(ctx, rest, binary_ops!(std::cmp::PartialOrd::gt))
-    }
-    intern_set_func!(ctx, gt, ">");
-
-    fn ge(ctx: &mut TulispContext, rest: &TulispObject) -> Result<TulispObject, Error> {
-        reduce_with(ctx, rest, binary_ops!(std::cmp::PartialOrd::ge))
-    }
-    intern_set_func!(ctx, ge, ">=");
-
-    fn lt(ctx: &mut TulispContext, rest: &TulispObject) -> Result<TulispObject, Error> {
-        reduce_with(ctx, rest, binary_ops!(std::cmp::PartialOrd::lt))
-    }
-    intern_set_func!(ctx, lt, "<");
-
-    fn le(ctx: &mut TulispContext, args: &TulispObject) -> Result<TulispObject, Error> {
-        reduce_with(ctx, args, binary_ops!(std::cmp::PartialOrd::le))
-    }
-    intern_set_func!(ctx, le, "<=");
-
-    fn max(ctx: &mut TulispContext, rest: &TulispObject) -> Result<TulispObject, Error> {
-        reduce_with(ctx, rest, max_min_ops!(max))
-    }
-    intern_set_func!(ctx, max, "max");
-
-    fn min(ctx: &mut TulispContext, rest: &TulispObject) -> Result<TulispObject, Error> {
-        reduce_with(ctx, rest, max_min_ops!(min))
-    }
-    intern_set_func!(ctx, min, "min");
-
     #[crate_fn(add_func = "ctx")]
     fn equal(object1: TulispObject, object2: TulispObject) -> bool {
         object1.equal(&object2)
