+++ src/builtin/functions/mod.rs
@@ -51,6 +51,7 @@ macro_rules! intern_set_func {
 }
 
 mod arithmetic_operations;
+mod comparison_of_numbers;
 mod conditionals;
 mod functions;
 mod list_elements;
@@ -58,6 +59,7 @@ mod sequences;
 
 pub(crate) fn add(ctx: &mut TulispContext) {
     arithmetic_operations::add(ctx);
+    comparison_of_numbers::add(ctx);
     conditionals::add(ctx);
     functions::add(ctx);
     sequences::add(ctx);
